requirements.txt
asgiref==3.4.1
astroid==2.9.0
awsebcli==3.20.3
backports.entry-points-selectable==1.1.0
bandit==1.7.4
botocore==1.23.54
cement==2.8.2
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.3
cycler==0.11.0
distlib==0.3.3
dj-database-url==0.5.0
Django==2.2
django-heroku==0.3.1
djangorestframework==3.13.1
filelock==3.3.1
fonttools==4.31.2
future==0.16.0
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
idna==3.3
isort==5.10.1
jmespath==0.10.0
joblib==1.1.0
kiwisolver==1.4.2
lazy-object-proxy==1.6.0
matplotlib==3.5.1
mccabe==0.6.1
mysql-connector-python==8.0.28
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pathspec==0.9.0
pbr==5.8.1
Pillow==8.4.0
platformdirs==2.4.0
protobuf==3.20.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pylint==2.12.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==3.0.7
pypiwin32==223
python-dateutil==2.8.2
pytz==2021.3
pywin32==303
PyYAML==5.4.1
requests==2.26.0
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
semantic-version==2.8.5
six==1.14.0
sklearn==0.0
smmap==5.0.0
sqlparse==0.4.2
stevedore==3.5.0
termcolor==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
utm==0.7.0
virtualenv==20.8.1
wcwidth==0.1.9
whitenoise==5.3.0
wrapt==1.13.3
